pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "baircondor"
version = "0.1.0"
description = "HTCondor job submission helper for BAIR lab GPU servers"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
]

[project.scripts]
baircondor = "baircondor.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["baircondor*"]
